Move cairo runner to runners package
